[MOL-13533][JS] move interfaces and types to types.ts file
