+++ src/myinfo/fake/index.ts
@@ -1,6 +1,7 @@
 import * as factory from "./factory/person-basic-factory";
 export { factory };
-export { FakeMyInfoHelper as Helper, IFakeMyInfoHelper as IHelper, MockParams } from "./fake-helper";
+export { FakeMyInfoHelper as Helper, IFakeMyInfoHelper as IHelper } from "./fake-helper";
 export { FakeMyInfoRequest as Request } from "./fake-request";
 export * from "./profiles/fake-profiles";
 export { ProfileArchetype } from "./profiles/fake-profile";
+export { MockParams } from "./types";
